# Swagger Docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Pweb-Backend-API",
        "version": "1.0",
        "title": "Swagger API",
        "contact": {
            "name": "Facultad Informática"
        }
    },
    "host": "localhost:8085",
    "basePath": "/",
...
    "paths": {
...
        "/api/v1/clientes/{id}": {
            "get": {
                "tags": [
                    "cliente-controller"
                ],
                "summary": "getUserById",
                "operationId": "getUserByIdUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ClienteDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "cliente-controller"
                ],
                "summary": "delete",
                "operationId": "deleteUsingDELETE_3",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/v1/roles/users/{userId}": {
            "get": {
                "tags": [
                    "role-controller"
                ],
                "summary": "geRoleByUserId",
                "operationId": "geRoleByUserIdUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "userId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RoleDto"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "summary": "getUsers",
                "operationId": "getUsersUsingGET",
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/UserDto"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "user-controller"
                ],
                "summary": "create",
                "operationId": "createUsingPOST_9",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "user",
                        "description": "user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "user-controller"
                ],
                "summary": "update",
                "operationId": "updateUsingPUT_9",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "user",
                        "description": "user",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/users/username/{username}": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "summary": "getUserByUsername",
                "operationId": "getUserByUsernameUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "username",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "summary": "getUserById",
                "operationId": "getUserByIdUsingGET_1",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "user-controller"
                ],
                "summary": "delete",
                "operationId": "deleteUsingDELETE_9",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ClienteDto": {
            "type": "object",
            "properties": {
                "annosServicio": {
                    "type": "integer",
                    "format": "int32"
                },
                "annos_suplidos": {
                    "type": "integer",
                    "format": "int32"
                },
                "antiguedad": {
                    "type": "integer",
                    "format": "int32"
                },
                "clientePriorizado": {
                    "type": "boolean"
                },
                "direccionCliente": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "faxCliente": {
                    "type": "string"
                },
                "nombCliente": {
                    "type": "string"
                },
                "paisCliente": {
                    "type": "string"
                },
                "tarifaAlmacenamiento": {
                    "type": "number",
                    "format": "double"
                },
                "tarifaRefrigeracion": {
                    "type": "number",
                    "format": "double"
                },
                "tarifaTransporte": {
                    "type": "number",
                    "format": "double"
                },
                "telefCliente": {
                    "type": "string"
                },
                "tipo_cliente": {
                    "type": "string"
                }
            },
            "title": "ClienteDto"
        },
...
        "RoleDto": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "roleName": {
                    "type": "string"
                }
            },
            "title": "RoleDto"
        },
...
        "UserDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identification": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RoleDto"
                    }
                },
                "username": {
                    "type": "string"
                }
            },
            "title": "UserDto"
        }
    }
}